backend_django/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.10
django-environ==0.12.0
dj-database-url==3.0.1
gunicorn==23.0.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
